tests/wpt/tests/encoding-detection/ar-windows-1256-late.tentative.html

<!doctype html>
<meta charset="utf-8">
<title>ar windows-1256 late</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
setup({explicit_done:true});
window.onmessage = function(e) {
    test(function() {
        assert_equals(e.data, "windows-1256", 'Expected windows-1256');
    }, "Check detection result");
    w.close();
    done();
};
var w = window.open("support/ar-windows-1256-late.sub.html");
</script>
